Use key_url instead of deprecated SKS keyserver
From https://sks-keyservers.net/:
This service is deprecated. This means it is no longer maintained, and
new HKPS certificates will not be issued. Service reliability should not
be expected.

Update 2021-06-21: Due to even more GDPR takedown requests, the DNS
records for the pool will no longer be provided at all.